[docs] refactor RouteUrl component
