+++ docs/ui/components/RouteUrl/RuntimePopup.tsx
@@ -1,15 +1,18 @@
-import { css } from '@emotion/react';
-import { ExpoGoLogo, shadows, theme, typography } from '@expo/styleguide';
-import { breakpoints, spacing } from '@expo/styleguide-base';
+import { ExpoGoLogo, mergeClasses } from '@expo/styleguide';
 import { ChevronDownIcon, Monitor01DuotoneIcon, Phone01DuotoneIcon } from '@expo/styleguide-icons';
 import { useEffect, useState } from 'react';
 
-type PopupActionProps<T extends string> = {
+type RuntimePopupProps<T extends string> = {
   items: { name: string; id: T }[];
   selected: string;
   onSelect: (value: T) => void;
 };
-export function RuntimePopup<T extends string>({ items, selected, onSelect }: PopupActionProps<T>) {
+
+export function RuntimePopup<T extends string>({
+  items,
+  selected,
+  onSelect,
+}: RuntimePopupProps<T>) {
   const Icon = [ExpoGoLogo, Phone01DuotoneIcon, Monitor01DuotoneIcon][
     items.findIndex(item => item.id === selected)
   ];
@@ -24,8 +27,12 @@ export function RuntimePopup<T extends string>({ items, selected, onSelect }: Po
       <select
         aria-label="Runtime URL format selector"
         title="Select runtime URL format"
-        css={selectStyle}
-        className="focus-visible:-outline-offset-2 border-0 rounded-none border-l border-l-default h-10 leading-10 px-10 hocus:bg-subtle hocus:shadow-none"
+        className={mergeClasses(
+          'flex items-center justify-center text-default h-10 px-10 m-0 min-w-[100px] rounded-none border-l indent-0 border-l-default shadow-xs appearance-none bg-default text-sm',
+          'hocus:bg-subtle hocus:shadow-none',
+          'focus-visible:-outline-offset-2',
+          'max-md-gutters:indent-[-9999px] max-md-gutters:min-w-[unset] max-md-gutters:max-w-[60px] max-md-gutters:px-6'
+        )}
         value={selected}
         onChange={e => {
           onSelect(e.target.value as T);
@@ -37,9 +44,7 @@ export function RuntimePopup<T extends string>({ items, selected, onSelect }: Po
         ))}
       </select>
       {isLoaded && (
-        <div
-          style={{ lineHeight: 1.3 }}
-          className="absolute inset-x-2.5 inset-y-0 flex items-center justify-between gap-2 text-icon-secondary pointer-events-none select-none">
+        <div className="absolute inset-x-3 inset-y-0 flex items-center justify-between gap-2 text-icon-secondary pointer-events-none select-none">
           <Icon className={ICON_CLASSES} />
           <ChevronDownIcon className="icon-xs text-icon-secondary pointer-events-none" />
         </div>
@@ -49,35 +54,3 @@ export function RuntimePopup<T extends string>({ items, selected, onSelect }: Po
 }
 
 const ICON_CLASSES = 'icon-sm text-icon-secondary pointer-events-none inline-block';
-
-const selectStyle = css`
-  ${typography.fontSizes[14]}
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: ${theme.text.default};
-  line-height: 1.3;
-  padding: 0 ${spacing[8]}px;
-  color: ${theme.text.default};
-  text-indent: 0;
-
-  box-shadow: ${shadows.xs};
-  -moz-appearance: none;
-  -webkit-appearance: none;
-  appearance: none;
-  background-color: ${theme.background.default};
-  cursor: pointer;
-
-  :hover {
-    background-color: ${theme.background.element};
-  }
-
-  :focus-visible {
-    background-color: ${theme.background.element};
-  }
-
-  @media screen and (max-width: ${(breakpoints.medium + breakpoints.large) / 2}px) {
-    padding: 0 0;
-    text-indent: -9999px;
-  }
-`;
